// ANGULAR/ANGULAR/src/app/Shared/Model/commonddl.model.ts
export class FilterDDlModel {
  FilterFor: string;
  Value: any;
  FilterFrom: string;
}

export class DdlItemModel {
  Value: any;
  Text: string;
}