// ANGULAR/ANGULAR/src/app/Shared/Enum/user-type.enum.ts
export enum UserTypeEnum {
  SADM = "SADM",
  ADM = "ADM",
  CMOO = "CMOO",
  DPTO = "DPTO",
  CMOS = "CMOS",
  DCOM = "DCOM",
  COLL = "COLL",
  BDO = "BDO",
  TEHDR = "TEHDR",
  MNSTR = "MNSTR",
  TLO = "TLO",
  DPTS = "DPTS",
  MLAMP = "MLAMP",
  DPTOFCA = "DPTOFCA",
  DPTOFCS = "DPTOFCS",
  // code by ravi
  DLO = "DLO",
  DLS = "DLS",
  BLO = "BLO",
  // code by ravi
}

export enum ConfigTypeEnum {
  Self = "self",
  Assign = "assign",
}
export enum ConfigUserModeEnum {
  UserId = "userid",
  UserType = "usertype",
}

export enum SessionTimeOut {
  TimeInMinul = 30,
}